requirements.txt
# 第一阶段依赖
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
anthropic>=0.18.0
pydantic>=2.0.0

# 第二阶段依赖
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
APScheduler>=3.10.0
Jinja2>=3.1.0
resend>=0.8.0

# 测试依赖
pytest>=7.4.0
pytest-cov>=4.1.0

# 开发工具
black>=23.12.0
ruff>=0.1.0
mypy>=1.0.0
